VoronoiGraph

Documentation for VoronoiGraph.

VoronoiGraph.adjacency

VoronoiGraph.boundary_area

VoronoiGraph.boundary_area_vrep

VoronoiGraph.descent

VoronoiGraph.explore

VoronoiGraph.mc_integrate

VoronoiGraph.mc_volume

VoronoiGraph.mc_volumes

VoronoiGraph.mc_volumes

VoronoiGraph.neighbors

VoronoiGraph.randray

VoronoiGraph.raycast

VoronoiGraph.raycast

VoronoiGraph.raycast

VoronoiGraph.transformation

VoronoiGraph.volumes

VoronoiGraph.voronoi

VoronoiGraph.voronoi_random

VoronoiGraph.walk

VoronoiGraph.walkray

VoronoiGraph.adjacency — Method

given vertices in generator-coordinates, collect the verts belonging to generator pairs, i.e. boundary vertices

VoronoiGraph.boundary_area — Method

given two generators g1, g2, vertices of their common boundary (in generator representation) and the list of all generators, compute the boundary volume. It works by constructing a polytope from halfspaces between the generators of the boundary vertices. We use an affine transformation to get rid of the dimension in direction g1->g2

VoronoiGraph.boundary_area_vrep — Method

similar to boundary_area, however uses a vector representation and is slower

VoronoiGraph.descent — Function

starting at given points, run the ray shooting descent to find vertices

VoronoiGraph.explore — Method

BFS of vertices starting from S0

VoronoiGraph.mc_integrate — Function

mc_integrate(f::Function, i::Int, xs::Points, nmc=1000, nmc2=1000, searcher=Raycast(xs))

Integrate function f over cell i and its boundary using nmc rays per cell and nmc2 points per ray for the volume integral.

Returns:

Vf::Real: the volume integral of f

Af::SparseVector: Af[j] is the surface integral of f over the intersection between cells i and j.

V::Real: the volume of cell i

A::SparseVector: A[j] is the surface area of the intersection between cells i and j.

VoronoiGraph.mc_volume — Function

mc_volume(i, xs, nmc, searcher)

Estimate the area and volume of the i-th Voronoi cell from the Voronoi Diagram generated by xs by a Monte Carlo estimate from random rays

VoronoiGraph.mc_volumes — Function

mc_volumes(xs::Points, nmc=1000)

Estimate the areas and volumes of the Voronoi Cells generated by xs using nmc Monte Carlo samples.

Returns

SparseMatrix: the areas of the common boundaries of two cells

Vector: the volumes of each cell

VoronoiGraph.mc_volumes — Function

mc_volumes(sig::Vertices, xs::Points, nmc=1000)

In the case when the simplicial complex is already known this information can be used to speed up the Monte-Carlo sampling by restricting the search space

VoronoiGraph.neighbors — Method

neighbors(sig::Vertices) --> Dict{Int, Vector{Int}}

Compute the neighbors of all generators.

VoronoiGraph.randray — Method

generate a random ray orthogonal to the subspace spanned by the given points

VoronoiGraph.raycast — Method

shooting a ray in the given direction, find the next connecting point. This variant (by Poliaski, Pokorny) uses a binary search

VoronoiGraph.raycast — Method

Shooting a ray in the given direction, find the next connecting point. This variant uses an iterative NN search

VoronoiGraph.raycast — Method

shooting a ray in the given direction, find the next connecting point. This is the bruteforce variant, using a linear search to find the closest point

VoronoiGraph.transformation — Method

affine transformation rotatinig and translating such that the boundary is aligned with the first dimension. A->B will be mapped to const*[1,0,0,...] and (A+B)/2 to [0,0,...]

VoronoiGraph.volumes — Method

build the connectivity matrix for the SQRA from adjacency and boundary information

VoronoiGraph.voronoi — Function

construct the voronoi diagram from x through breadth-first search

VoronoiGraph.voronoi_random — Function

construct a (partial) voronoi diagram from x through a random walk

VoronoiGraph.walk — Method

starting at vertices, walk nsteps along the voronoi graph to find new vertices

VoronoiGraph.walkray — Method

find the vertex connected to v by moving away from its i-th generator