VoronoiGraph

Documentation for VoronoiGraph.

VoronoiGraph.adjacency

VoronoiGraph.area_volume

VoronoiGraph.boundary_area

VoronoiGraph.boundary_area_vrep

VoronoiGraph.descent

VoronoiGraph.explore

VoronoiGraph.randray

VoronoiGraph.raycast

VoronoiGraph.raycast

VoronoiGraph.raycast

VoronoiGraph.transformation

VoronoiGraph.voronoi

VoronoiGraph.voronoi_random

VoronoiGraph.walk

VoronoiGraph.walkray

VoronoiGraph.walkray

VoronoiGraph.adjacency — Method

given vertices in generator-coordinates, collect the verts belonging to generator pairs, i.e. boundary vertices

VoronoiGraph.area_volume — Method

build the connectivity matrix for the SQRA from adjacency and boundary information

VoronoiGraph.boundary_area — Method

given two generators g1, g2, vertices of their common boundary (in generator representation) and the list of all generators, compute the boundary volume. It works by constructing a polytope from halfspaces between the generators of the boundary vertices. We use an affine transformation to get rid of the dimension in direction g1->g2

VoronoiGraph.boundary_area_vrep — Method

similar to boundary_area, however uses a vector representation and is slower

VoronoiGraph.descent — Function

starting at given points, run the ray shooting descent to find vertices

VoronoiGraph.explore — Method

BFS of vertices starting from S0

VoronoiGraph.randray — Method

generate a random ray orthogonal to the subspace spanned by the given points

VoronoiGraph.raycast — Method

shooting a ray in the given direction, find the next connecting point. This variant (by Poliaski, Pokorny) uses a binary search

VoronoiGraph.raycast — Method

Shooting a ray in the given direction, find the next connecting point. This variant uses an iterative NN search

VoronoiGraph.raycast — Method

shooting a ray in the given direction, find the next connecting point. This is the bruteforce variant, using a linear search to find the closest point

VoronoiGraph.transformation — Method

affine transformation rotatinig and translating such that the boundary is aligned with the first dimension. A->B will be mapped to const*[1,0,0,...] and (A+B)/2 to [0,0,...]

VoronoiGraph.voronoi — Method

construct the voronoi diagram from x through breadth-first search

VoronoiGraph.voronoi_random — Function

construct a (partial) voronoi diagram from x through a random walk

VoronoiGraph.walk — Method

starting at vertices, walk nsteps along the voronoi graph to find new vertices

VoronoiGraph.walkray — Method

starting at vertex (v,r), return a random adjacent vertex

VoronoiGraph.walkray — Method

find the vertex connected to v by moving away from its i-th generator